Move role dependencies from bootstrap to RoleServiceProvider
