+++ bootstrap.php
@@ -326,110 +326,9 @@ $container->addServiceProvider(new OrganizerJSONLDServiceProvider());
 $container->addServiceProvider(new OrganizerCommandHandlerProvider());
 
 /** Roles */
-$container->addShared(
-    'role_iri_generator',
-    fn () => new CallableIriGenerator(
-        fn ($roleId) => $container->get('config')['url'] . '/roles/' . $roleId
-    )
-);
+$container->addServiceProvider(new \CultuurNet\UDB3\Role\RoleServiceProvider());
 
-$container->addShared(
-    'role_store',
-    fn () => $container->get('event_store_factory')(AggregateType::role())
-);
-
-$container->addShared(
-    'real_role_repository',
-    fn () => new \CultuurNet\UDB3\Role\RoleRepository(
-        $container->get('role_store'),
-        $container->get(EventBus::class),
-    )
-);
-
-$container->addShared(
-    'role_read_repository',
-    fn () => new \CultuurNet\UDB3\ReadModel\BroadcastingDocumentRepositoryDecorator(
-        new \CultuurNet\UDB3\Doctrine\ReadModel\CacheDocumentRepository(
-            $container->get('cache')('role_detail')
-        ),
-        $container->get(EventBus::class),
-        new \CultuurNet\UDB3\Role\ReadModel\Detail\EventFactory()
-    )
-);
-
-$container->addShared(
-    'user_roles_repository',
-    fn () => new \CultuurNet\UDB3\Doctrine\ReadModel\CacheDocumentRepository($container->get('cache')('user_roles'))
-);
-
-/**
- * @todo move this to a class.
- */
 const ROLE_SEARCH_V3_REPOSITORY_TABLE_NAME = 'roles_search_v3';
-
-$container->addShared(
-    'role_search_v3_repository',
-    fn () => new \CultuurNet\UDB3\Role\ReadModel\Search\Doctrine\DBALRepository(
-        $container->get('dbal_connection'),
-        new StringLiteral(ROLE_SEARCH_V3_REPOSITORY_TABLE_NAME)
-    )
-);
-
-$container->addShared(
-    'role_search_v3_projector',
-    fn () => new \CultuurNet\UDB3\Role\ReadModel\Search\Projector($container->get('role_search_v3_repository'))
-);
-
-$container->addShared(
-    'role_detail_projector',
-    fn () => new \CultuurNet\UDB3\Role\ReadModel\Detail\Projector($container->get('role_read_repository'))
-);
-
-$container->addShared(
-    'user_roles_projector',
-    fn () => new \CultuurNet\UDB3\Role\ReadModel\Users\UserRolesProjector(
-        $container->get('user_roles_repository'),
-        $container->get('role_read_repository'),
-        $container->get('role_users_read_repository'),
-    )
-);
-
-$container->addShared(
-    'role_labels_read_repository',
-    fn () => new \CultuurNet\UDB3\Doctrine\ReadModel\CacheDocumentRepository($container->get('cache')('role_labels'))
-);
-
-$container->addShared(
-    'role_labels_projector',
-    fn () => new \CultuurNet\UDB3\Role\ReadModel\Labels\RoleLabelsProjector(
-        $container->get('role_labels_read_repository'),
-        $container->get('labels.json_read_repository'),
-        $container->get('label_roles_read_repository'),
-    )
-);
-
-$container->addShared(
-    'label_roles_read_repository',
-    fn () => new \CultuurNet\UDB3\Doctrine\ReadModel\CacheDocumentRepository($container->get('cache')('label_roles'))
-);
-
-$container->addShared(
-    'label_roles_projector',
-    fn () => new \CultuurNet\UDB3\Role\ReadModel\Labels\LabelRolesProjector($container->get('label_roles_read_repository'))
-);
-
-$container->addShared(
-    'role_users_read_repository',
-    fn () => new \CultuurNet\UDB3\Doctrine\ReadModel\CacheDocumentRepository($container->get('cache')('role_users'))
-);
-
-$container->addShared(
-    'role_users_projector',
-    fn () => new \CultuurNet\UDB3\Role\ReadModel\Users\RoleUsersProjector(
-        $container->get('role_users_read_repository'),
-        $container->get(Auth0UserIdentityResolver::class),
-    )
-);
 
 $app['event_export_notification_mail_factory'] = $app->share(
     function ($app) {
